+++ src/components/main/index.js
@@ -1,0 +1,2 @@
+import Main from './main.js'
+export { Main };
